requirements.txt
psycopg2-binary

daphne
django==3.1
djangorestframework==3.11.0
django-split-settings
celery
pillow

django-webtest
model-mommy
